requirements.txt
click==8.2.1
dlib==20.0.0
face-recognition==1.3.0
face_recognition_models==0.3.0
numpy==2.2.6
opencv-python-headless==4.11.0.86
pandas==2.3.0
pillow==11.2.1
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
tzdata==2025.2
